clearly divide the work of agent class & long memory stream
