+++ prompt/memory_stream/common_example.json
