+++ prompt/memory_stream/common_prompt.json
